media-source/MediaSource-endOfStream-step2.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>insertAdjacentHTML</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>

<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }        
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    GET(FILE, function(uInt8Array) {
        var data = new Blob([uInt8Array], {type: 'video/webm'});
        var chunkNum = 5;
        var chunkSize = Math.ceil(data.size / chunkNum);
        var chunkIdx = 0;
        var mediaSource = new MediaSource();

        mediaSource.addEventListener('sourceopen', function(e){
            //playback 1st chunk
            var sourceBuffer = this.addSourceBuffer('video/webm; codecs="vorbis,vp8"');
            var startByte = chunkSize * chunkIdx;
            var chunk = data.slice(startByte, startByte + chunkSize);
            var reader = new FileReader();
            reader.onload = function(e) {
                var data = new Uint8Array(e.target.result);
                sourceBuffer.appendBuffer(data);
                chunkIdx++;
            };
            reader.readAsArrayBuffer(chunk);

            sourceBuffer.addEventListener('updatestart', function(e) {
                console.log('updatestart: ' + this.updating);
                t.step(function() {
                    assert_equals(sourceBuffer.updating, true);
                    assert_throws({name : 'InvalidStateError'},
                                  function() {
                                      mediaSource.endOfStream();
                                  });
                });
                t.done();
            });
            sourceBuffer.addEventListener('update', function(e) {                
                
            });
        }, false);
        video.src = window.URL.createObjectURL(mediaSource);
    });
}, "MediaSource#endOfStream : 2. If the updating attribute equals true on any SourceBuffer in sourceBuffers, then throw an INVALID_STATE_ERR exception and abort these steps.");

</script>
</body>
</html>
